Max Updates v1.51

+++ ZodiaClash/ZodiaClash/enginecore.h
@@ -45,7 +45,7 @@ public:
 
 private:
 
-	const uint64_t m_initialTime = std::chrono::time_point_cast<std::chrono::microseconds>(std::chrono::steady_clock::now()).time_since_epoch().count();
+	const uint64_t m_initialTime = static_cast<uint64_t>(std::chrono::time_point_cast<std::chrono::microseconds>(std::chrono::steady_clock::now()).time_since_epoch().count());
 		
 	uint64_t m_previousTime;
